refactor: 상품 수정 요청 필드 수정
상품 수정 요청 필드 검증 조건 수정

CHZZ-176
